WebApp/requirements.txt
Flask==3.0.3
torch==2.3.1
torchvision==0.18.1
pillow==10.4.0
numpy==1.26.4
gunicorn==22.0.0